requirements.txt
aiogram==2.15
aiogram-media-group @ git+https://github.com/YulLeo/aiogram-media-group.git@8cc1f1de7cb87e184a4646f08355dad4f18cf8e3
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0
attrs==21.2.0
Babel==2.9.1
black==21.10b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
flake8==4.0.1
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
images2gif==1.0.1
iniconfig==1.1.1
isort==5.10.1
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
packaging==21.2
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.1
pytz==2021.3
regex==2021.11.2
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
yarl==1.7.2

minio~=7.1.1